RF: Move segmentation workflow to standalone module
